requirements-train.txt
torch
transformers
datasets
peft
trl
accelerate
bitsandbytes
safetensors
anthropic
gradio
llama-cpp-python

# torch>=2.3.0
# transformers>=4.51.0
# datasets>=2.20.0
# peft>=0.14.0
# trl>=0.9.6
# accelerate>=0.33.0
# bitsandbytes>=0.43.0
# safetensors>=0.4.3
# anthropic>=0.30.0
# gradio>=4.40.0
# llama-cpp-python>=0.2.90